MCP-1935: extracted common domain
