+++ api/build.gradle
@@ -12,5 +12,7 @@ dependencies {
 
     implementation 'io.micrometer:micrometer-core'
     implementation 'gov.va.starter:exceptions'
+
+    implementation project(':domain')
 }
