+++ controller/build.gradle
@@ -11,6 +11,7 @@ dependencies {
     implementation "org.springframework.boot:spring-boot-dependencies:${spring_boot_version}"
     implementation "org.springframework:spring-webmvc:${spring_framework_version}"
 
+    implementation project(':domain')
     implementation project(':api')
     implementation project(':service:provider')
     implementation project(':service:spi')
