+++ domain/src/main/java/gov/va/vro/model/AbdEvidence.java
@@ -1,5 +1,7 @@
-package gov.va.vro.service.spi.model;
+package gov.va.vro.model;
 
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
 import lombok.Setter;
@@ -8,6 +10,8 @@ import java.util.List;
 
 @Getter
 @Setter
+@JsonInclude(JsonInclude.Include.NON_NULL)
+@JsonIgnoreProperties(ignoreUnknown = true)
 public class AbdEvidence {
   private List<AbdMedication> medications;
   private List<AbdCondition> conditions;
