+++ domain/src/main/java/gov/va/vro/model/AbdProcedure.java
@@ -1,4 +1,4 @@
-package gov.va.vro.api.model;
+package gov.va.vro.model;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
