--- service/spi/src/main/java/gov/va/vro/service/spi/model/AbdBpMeasurement.java
@@ -1,15 +1,0 @@
-package gov.va.vro.service.spi.model;
-
-import lombok.Getter;
-import lombok.Setter;
-
-import java.math.BigDecimal;
-
-@Getter
-@Setter
-public class AbdBpMeasurement {
-  private String code;
-  private String display;
-  private String unit;
-  private BigDecimal value;
-}
